Remove common `handlers.Interface`
Closes #3662.
