+++ internal/handlers/registry/mysql.go
@@ -16,13 +16,12 @@ package registry
 
 import (
 	"github.com/FerretDB/FerretDB/internal/backends/mysql"
-	"github.com/FerretDB/FerretDB/internal/handlers"
 	handler "github.com/FerretDB/FerretDB/internal/handlers/sqlite"
 )
 
 // init registers "mysql" handler.
 func init() {
-	registry["mysql"] = func(opts *NewHandlerOpts) (handlers.Interface, CloseBackendFunc, error) {
+	registry["mysql"] = func(opts *NewHandlerOpts) (*handler.Handler, CloseBackendFunc, error) {
 		b, err := mysql.NewBackend(&mysql.NewBackendParams{
 			URI: opts.MySQLURL,
 			L:   opts.Logger.Named("mysql"),
