+++ internal/handlers/sqlite/msg_datasize.go
@@ -29,7 +29,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgDataSize implements HandlerInterface.
+// MsgDataSize implements `dataSize` command.
 func (h *Handler) MsgDataSize(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	document, err := msg.Document()
 	if err != nil {
