+++ internal/handlers/sqlite/msg_dbstats.go
@@ -28,7 +28,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgDBStats implements HandlerInterface.
+// MsgDBStats implements `dbStats` command.
 func (h *Handler) MsgDBStats(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	document, err := msg.Document()
 	if err != nil {
