+++ internal/handlers/sqlite/msg_debugerror.go
@@ -21,7 +21,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgDebugError implements HandlerInterface.
+// MsgDebugError implements `debugError` command.
 func (h *Handler) MsgDebugError(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	return commoncommands.MsgDebugError(ctx, msg)
 }
