+++ internal/handlers/sqlite/msg_delete.go
@@ -31,7 +31,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgDelete implements HandlerInterface.
+// MsgDelete implements `delete` command.
 func (h *Handler) MsgDelete(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	document, err := msg.Document()
 	if err != nil {
