+++ internal/handlers/sqlite/msg_distinct.go
@@ -28,7 +28,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgDistinct implements HandlerInterface.
+// MsgDistinct implements `distinct` command.
 func (h *Handler) MsgDistinct(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	document, err := msg.Document()
 	if err != nil {
