+++ internal/handlers/sqlite/msg_dropindexes.go
@@ -30,7 +30,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgDropIndexes implements HandlerInterface.
+// MsgDropIndexes implements `dropIndexes` command.
 func (h *Handler) MsgDropIndexes(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	document, err := msg.Document()
 	if err != nil {
