+++ internal/handlers/sqlite/msg_find.go
@@ -32,7 +32,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgFind implements HandlerInterface.
+// MsgFind implements `find` command.
 func (h *Handler) MsgFind(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	document, err := msg.Document()
 	if err != nil {
