+++ internal/handlers/sqlite/msg_getlog.go
@@ -33,7 +33,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgGetLog implements HandlerInterface.
+// MsgGetLog implements `getLog` command.
 func (h *Handler) MsgGetLog(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	document, err := msg.Document()
 	if err != nil {
