+++ internal/handlers/sqlite/msg_ismaster.go
@@ -23,7 +23,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgIsMaster implements HandlerInterface.
+// MsgIsMaster implements `isMaster` command.
 func (h *Handler) MsgIsMaster(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	doc, err := msg.Document()
 	if err != nil {
