+++ internal/handlers/sqlite/msg_killcursors.go
@@ -21,7 +21,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgKillCursors implements handlers.Interface.
+// MsgKillCursors implements `killCursors` command.
 func (h *Handler) MsgKillCursors(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	return common.KillCursors(ctx, msg, h.cursors)
 }
