+++ internal/handlers/sqlite/msg_listdatabases.go
@@ -27,7 +27,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgListDatabases implements HandlerInterface.
+// MsgListDatabases implements `listDatabases` command.
 func (h *Handler) MsgListDatabases(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	document, err := msg.Document()
 	if err != nil {
