+++ internal/handlers/sqlite/msg_listindexes.go
@@ -27,7 +27,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgListIndexes implements HandlerInterface.
+// MsgListIndexes implements `listIndexes` command.
 func (h *Handler) MsgListIndexes(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	document, err := msg.Document()
 	if err != nil {
