+++ internal/handlers/sqlite/msg_saslstart.go
@@ -24,7 +24,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgSASLStart implements HandlerInterface.
+// MsgSASLStart implements `saslStart` command.
 func (h *Handler) MsgSASLStart(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	document, err := msg.Document()
 	if err != nil {
