+++ internal/handlers/sqlite/msg_update.go
@@ -31,7 +31,7 @@ import (
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
-// MsgUpdate implements HandlerInterface.
+// MsgUpdate implements `update` command.
 func (h *Handler) MsgUpdate(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	document, err := msg.Document()
 	if err != nil {
